requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pandas==1.3.5
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3